+++ src/http/response.hpp
@@ -14,9 +14,6 @@
 #include "body_stream.hpp"
 #include "http_tools.hpp"
 
-// The minimum size for a body to be compressed
-#define MIN_COMPRESSION_SIZE 750
-
 namespace http {
 
     class Response {