+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.23.0
 - Justified CLI table in README (#240)
 - Updated Performance section in README to include what version was tested
+- Add config controls for keep-alive connections & minimum response body size for compression (#237)
 
 ## v0.22.2
 - Removed update.ps1 and update.sh update scripts